+++ src/main/java/com/chunkrsvp/cli/input/InputAction.java
@@ -1,0 +1,5 @@
+package com.chunkrsvp.cli.input;
+
+public enum InputAction {
+    PAUSE_TOGGLE, SPEED_UP, SPEED_DOWN, REWIND, EXIT, NONE
+}